Rename adlrtool to adlr-cli
