+++ adlr-cli/cmd/about_licenses.go
