+++ adlr-cli/cmd/evaluate.go
