+++ adlr-cli/cmd/root.go
