+++ adlr-cli/license.lock
